remove old thank you flow (#2752)
* remove old thank you flow

* remove unnecessary new-flow directory

* remove additional dead code
