--- support-frontend/assets/components/spreadTheWord/spreadTheWord.scss
@@ -1,20 +1,0 @@
-.component-spread-the-word {
-  padding: 0 10px ($gu-v-spacing * 2);
-  margin: 0 -10px;
-  border-bottom: 1px solid gu-colour(neutral-86);
-}
-
-.component-spread-the-word__title {
-  font-family: $gu-text-egyptian-web;
-  letter-spacing: -0.5px;
-  font-weight: 900;
-  font-size: 17px;
-  margin: ($gu-v-spacing / 3) 0;
-}
-
-.component-spread-the-word__description {
-  font-family: $gu-text-egyptian-web;
-  letter-spacing: -0.5px;
-  margin-bottom: $gu-v-spacing;
-  line-height: 20px;
-}
