+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/ContributionThankYouSupportReminder.jsx
@@ -18,7 +18,7 @@ import SvgClock from './components/SvgClock';
 import styles from './styles';
 import {
   OPHAN_COMPONENT_ID_SET_REMINDER,
-} from '../utils/ophan';
+} from './utils/ophan';
 import { trackComponentClick } from 'helpers/tracking/behaviour';
 import { privacyLink } from 'helpers/legal';
