+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/components/SvgShare.jsx
