--- support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/old-flow/ContributionThankYou.scss
@@ -1,132 +1,0 @@
-@import './ContributionThankYouSetPassword';
-
-.thank-you__container {
-  display: flex;
-  flex-flow: column-reverse nowrap;
-  justify-content: flex-start;
-  margin: auto;
-
-  @include mq($from: tablet) {
-    flex-flow: row nowrap;
-    justify-content: space-around;
-  }
-
-  @include mq($from: leftCol) {
-    justify-content: center;
-  }
-
-  .gu-content__blurb.gu-content__blurb--thank-you {
-    display: block;
-    border-top: none;
-
-    @include mq($from: tablet) {
-      border-top: 10px solid gu-colour(neutral-7);
-    }
-  }
-
-  .gu-content__blurb-header {
-    margin: 0;
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    font-weight: 900;
-    font-size: 17px;
-    line-height: 22px;
-    padding-bottom: $gu-v-spacing * 2;
-
-    @include mq($from: tablet) {
-      margin: 0 -10px;
-      border: 0;
-      left: 10px;
-      padding: $gu-v-spacing/2 $gu-h-spacing/2 $gu-v-spacing $gu-h-spacing/2;
-      width: 100%;
-      font-size: 36px;
-      line-height: 1.05;
-      font-family: $gu-headline;
-      font-weight: bold;
-      color: gu-colour(neutral-7);
-    }
-  }
-
-  .gu-content__blurb-header-break {
-    display: none;
-
-    @include mq($from: tablet) {
-      display: block;
-    }
-  }
-
-  .gu-content__blurb-blurb {
-    display: none;
-
-    @include mq($from: tablet) {
-      display: block;
-      border-bottom: 1px solid gu-colour(neutral-7);
-    }
-  }
-
-  .gu-content__return-link {
-    padding: 20px 10px;
-    margin: 0 -10px;
-  }
-
-  .component-marketing-consent.component-marketing-consent--newsletter {
-    border-bottom: 1px solid gu-colour(neutral-86);
-  }
-
-  .component-marketing-consent__message {
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    margin-bottom: $gu-v-spacing;
-    line-height: 20px;
-  }
-
-  .component-marketing-consent__title {
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    font-weight: 900;
-    font-size: 17px;
-    margin: ($gu-v-spacing / 3) 0;
-    line-height: 22px;
-  }
-
-  .contribution-thank-you-block {
-    border-bottom: 1px solid gu-colour(neutral-86);
-    font-weight: 300;
-    margin: 0 -10px;
-    padding: 0 10px ($gu-v-spacing * 2);
-  }
-
-  .contribution-thank-you-block__title {
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    font-weight: 900;
-    font-size: 17px;
-    margin: ($gu-v-spacing / 3) 0;
-    line-height: 22px;
-  }
-
-  .contribution-thank-you-block__message {
-    font-family: $gu-text-egyptian-web;
-    letter-spacing: -0.5px;
-    margin-bottom: $gu-v-spacing;
-    line-height: 20px;
-  }
-
-  .contribution-thank-you-block--reminders {
-    .component-button {
-      margin-top: $gu-v-spacing;
-    }
-
-    .component-radio-input {
-      margin-bottom: $gu-v-spacing;
-
-      &:first-of-type {
-        margin-top: $gu-v-spacing / 2;
-      }
-    }
-
-    .component-radio-input__text {
-      font-weight: normal;
-    }
-  }
-}
