--- support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/old-flow/ContributionThankYouSetPassword.scss
@@ -1,35 +1,0 @@
-.set-password__message {
-  font-family: $gu-text-egyptian-web;
-  letter-spacing: -0.5px;
-  margin-bottom: $gu-v-spacing * 2;
-  line-height: 20px;
-}
-
-.set-password__message {
-  margin-bottom: 0;
-}
-
-.set-password__form {
-  padding-bottom: 30px;
-}
-
-.form__submit.form__submit--create-account {
-  padding-bottom: $gu-v-spacing;
-}
-
-.set-password__title {
-  font-family: $gu-text-egyptian-web;
-  letter-spacing: -0.5px;
-  font-weight: 900;
-  font-size: 17px;
-  margin: ($gu-v-spacing / 3) 0;
-}
-
-.form__field.form__field--contribution-email {
-  margin-top: $gu-v-spacing * 2;
-}
-
-.component-button--create-account {
-  margin-top: $gu-v-spacing * 2;
-  margin-bottom: $gu-v-spacing;
-}
