+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/styles.js
