+++ support-frontend/assets/pages/contributions-landing/contributionsLanding.scss
@@ -8,9 +8,7 @@
 @import '~components/generalErrorMessage/generalErrorMessage';
 @import '~components/button/button';
 @import '~components/marketingConsent/marketingConsent';
-@import '~components/spreadTheWord/spreadTheWord';
 @import '~components/socialShare/socialShare';
-@import '~pages/contributions-landing/components/ContributionThankYou/old-flow/ContributionThankYou';
 @import '~pages/contributions-landing/components/StripePaymentRequestButton/StripePaymentRequestButton';
 
 $content-left-margin-leftCol: calc(50% - 480px);
