+++ support-frontend/assets/pages/paypal-error/payPalError.scss
@@ -10,7 +10,6 @@
 @import '~components/headers/header/header';
 @import '~components/pageSection/pageSection';
 @import '~components/questionsContact/questionsContact';
-@import '~components/spreadTheWord/spreadTheWord';
 @import '~components/socialShare/socialShare';
